--- src/main/java/app/majime/lims/manufacturer/usecase/GetManufacturer.java
@@ -1,12 +1,0 @@
-package app.majime.lims.manufacturer.usecase;
-
-public interface GetManufacturer {
-
-    Long getId();
-
-    String getName();
-    char getDeleted();
-
-    Long getAddress_id();
-
-}
