--- src/main/java/app/majime/lims/manufacturer/usecase/SetManufacturer.java
@@ -1,12 +1,0 @@
-package app.majime.lims.manufacturer.usecase;
-
-public interface SetManufacturer {
-
-    void setId(Long id);
-
-    void setName(String name);
-    void setDeleted(char deleted);
-
-    void address_id(Long address_id);
-
-}
