+++ src/main/java/app/majime/lims/result/ResultStatus.java
@@ -1,0 +1,7 @@
+package app.majime.lims.result;
+
+enum ResultStatus {
+    NEW,
+    APPROVED,
+    DELETED
+}
