--- src/main/java/app/majime/lims/specification/MaterialRepository.java
@@ -1,12 +1,0 @@
-package app.majime.lims.specification;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface MaterialRepository extends CrudRepository<Material,Long>{
-    Optional<Material> findByName(String name);
-
-}
-
-
