+++ src/main/java/app/majime/lims/specification/SpecificationStatus.java
@@ -1,0 +1,7 @@
+package app.majime.lims.specification;
+
+enum SpecificationStatus {
+    NEW,
+    APPROVED,
+    DELETED
+}
