--- src/main/java/app/majime/lims/supplier/SupplierRepository.java
@@ -1,9 +1,0 @@
-package app.majime.lims.supplier;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface SupplierRepository extends CrudRepository<Supplier,Long> {
-    Optional<Supplier> findByName(String name);
-}
